refactor: move blocks into own dir
