+++ packages/react-tinacms-inline/src/blocks/index.ts
@@ -16,10 +16,8 @@ limitations under the License.
 
 */
 
-export * from './inline-form'
-export * from './inline-field'
-export * from './inline-field-text'
-export * from './blocks/inline-field-blocks'
-export * from './blocks/block'
-export * from './blocks/inline-block-field'
-export * from './blocks/inline-block-field-controls'
+export * from './inline-block-text'
+export * from './inline-field-blocks'
+export * from './inline-block-field-controls'
+export * from './inline-block-field'
+export * from './block'
