+++ packages/react-tinacms-inline/src/blocks/inline-block-field.tsx
@@ -17,8 +17,7 @@ limitations under the License.
 */
 
 import * as React from 'react'
-import { InlineField, InlineFieldRenderProps } from './inline-field'
-import { InlineTextFieldProps } from './inline-field-text'
+import { InlineField, InlineFieldRenderProps } from '../inline-field'
 import { useInlineBlock } from './inline-field-blocks'
 
 export interface BlockFieldProps {
@@ -34,23 +33,4 @@ export function BlockField({ name, children }: BlockFieldProps) {
   const block = useInlineBlock()
   const fieldName = `${block.name}.${name}`
   return <InlineField name={fieldName}>{children}</InlineField>
-}
-
-/**
- * InlineTextField
- */
-export interface BlockText {
-  name: string
-}
-export function BlockText({ name }: InlineTextFieldProps) {
-  return (
-    <BlockField name={name}>
-      {({ input, status }) => {
-        if (status === 'active') {
-          return <input type="text" {...input} />
-        }
-        return <>{input.value}</>
-      }}
-    </BlockField>
-  )
 }
